+++ crates/clairvoyance/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod tokens;
 pub mod uniswap;
 pub mod utils;
+pub mod monitor;
